Examination for the Certificate of Proficiency in English (ECPE) Item Responses

Examination for the Certificate of Proficiency in English (ECPE) Item Responses

items_ecpe

Format

A matrix consisting of 2922 rows and 28 columns. Thus, there are \(N = 2922\) subject responses to \(J = 28\) items.

The subjects answered the following assessment items:

item01

item02

item03

item04

item05

item06

item07

item08

item09

item11

item12

item13

item14

item15

item16

item17

item18

item19

item20

item21

item22

item23

item24

item25

item26

item27

item28

References

Data originated from:

Tatsuoka, C. (2002). Data analytic methods for latent partially ordered classification models. Journal of the Royal Statistical Society: Series C (Applied Statistics), 51(3), 337–350. https://doi.org/10.1111/1467-9876.00272

Tatsuoka, K. K. (1984), Analysis of errors in fraction addition and subtraction problems (Final Report for Grant No. NIE-G-81-0002). Urbana: University of Illinois, Computer-Based Education Research Laboratory (CERL).

Data used in:

Chen, Y., Culpepper, S. A., & Liang, F. (2020). A sparse latent class model for cognitive diagnosis. Psychometrika, 1–33. https://doi.org/10.1007/s11336-019-09693-2

Culpepper, S. A. (2019). Estimating the cognitive diagnosis \(Q\) matrix with expert knowledge: Application to the fraction-subtraction dataset. Psychometrika, 84(2), 333–357. https://doi.org/10.1007/s11336-018-9643-8

Culpepper, S. A., & Chen, Y. (2019). Development and application of an exploratory reduced reparameterized unified model. Journal of Educational and Behavioral Statistics, 44(1), 3–24. https://doi.org/10.3102/1076998618791306

Chen, Y., Culpepper, S. A., Chen, Y., & Douglas, J. (2018). Bayesian estimation of the dina q matrix. Psychometrika, 83(1), 89–108. https://doi.org/10.1007/s11336-017-9579-4